Split client and server code
