+++ packages/mdx/src/client/scrollycoding.tsx
@@ -1,23 +1,14 @@
-import { visitAsync, toJSX } from "./unist-utils"
-import { Node, Parent } from "unist"
 import React from "react"
 import {
-  EditorSpring,
   EditorProps,
   EditorStep,
 } from "@code-hike/mini-editor"
-import { mapEditor, Code } from "./code"
-import { reduceSteps } from "./code-files-reducer"
-import { extractStepsInfo } from "./steps"
+import { Code } from "./code"
 import {
   Scroller,
   Step as ScrollerStep,
 } from "@code-hike/scroller"
-import {
-  Preview,
-  PresetConfig,
-  getPresetConfig,
-} from "./preview"
+import { Preview, PresetConfig } from "./preview"
 
 export function Scrollycoding({
   children,
@@ -76,42 +67,4 @@ export function Scrollycoding({
       </div>
     </section>
   )
-}
-
-export async function transformScrollycodings(
-  tree: Node,
-  config: { theme: any }
-) {
-  await visitAsync(
-    tree,
-    "mdxJsxFlowElement",
-    async node => {
-      if (node.name === "CH.Scrollycoding") {
-        await transformScrollycoding(node, config)
-      }
-    }
-  )
-}
-async function transformScrollycoding(
-  node: Node,
-  { theme }: { theme: any }
-) {
-  const editorSteps = await extractStepsInfo(
-    node as Parent,
-    { theme },
-    "merge step with previous"
-  )
-
-  const presetConfig = await getPresetConfig(
-    (node as any).attributes
-  )
-
-  toJSX(node, {
-    props: {
-      codeConfig: { theme },
-      editorSteps: editorSteps,
-      presetConfig,
-    },
-    appendProps: true,
-  })
 }
