+++ packages/mdx/src/client/spotlight.scss
