+++ packages/mdx/src/client/spotlight.tsx
@@ -1,18 +1,10 @@
-import { visitAsync, toJSX } from "./unist-utils"
-import { Node, Parent } from "unist"
 import React from "react"
 import {
-  EditorSpring,
   EditorProps,
   EditorStep,
 } from "@code-hike/mini-editor"
-import { mapEditor, Code } from "./code"
-import { extractStepsInfo } from "./steps"
-import {
-  Preview,
-  PresetConfig,
-  getPresetConfig,
-} from "./preview"
+import { Code } from "./code"
+import { Preview, PresetConfig } from "./preview"
 
 export function Spotlight({
   children,
@@ -71,43 +63,4 @@ export function Spotlight({
       </div>
     </div>
   )
-}
-
-export async function transformSpotlights(
-  tree: Node,
-  config: { theme: any }
-) {
-  await visitAsync(
-    tree,
-    "mdxJsxFlowElement",
-    async node => {
-      if (node.name === "CH.Spotlight") {
-        await transformSpotlight(node, config)
-      }
-    }
-  )
-}
-
-async function transformSpotlight(
-  node: Node,
-  { theme }: { theme: any }
-) {
-  const editorSteps = await extractStepsInfo(
-    node as Parent,
-    { theme },
-    "merge steps with header"
-  )
-
-  const presetConfig = await getPresetConfig(
-    (node as any).attributes
-  )
-
-  toJSX(node, {
-    props: {
-      codeConfig: { theme },
-      editorSteps: editorSteps,
-      presetConfig,
-    },
-    appendProps: true,
-  })
 }
