+++ packages/mdx/src/plugin/annotations.ts
@@ -1,122 +1,8 @@
-import React from "react"
-import { CodeAnnotation } from "@code-hike/smooth-code"
-import { CodeLink } from "./links"
 import { Code, relativeToAbsolute } from "@code-hike/utils"
+import { CodeAnnotation } from "@code-hike/smooth-code"
 import { Node, Parent } from "unist"
 import { wrapChildren } from "./to-estree"
-
-export function Annotation() {
-  return "error: code hike remark plugin not running or annotation isn't at the right place"
-}
-
-function Box({
-  children,
-  data,
-  theme,
-}: {
-  data: any
-  children: React.ReactNode
-  theme: any
-}) {
-  const border =
-    typeof data === "string"
-      ? data
-      : theme.tokenColors.find((tc: any) =>
-          tc.scope?.includes("string")
-        )?.settings?.foreground || "yellow"
-  return (
-    <span style={{ outline: `2px solid ${border}` }}>
-      {children}
-    </span>
-  )
-}
-
-function Background({
-  children,
-  data,
-  style,
-  theme,
-}: {
-  data: string
-  children: React.ReactNode
-  style?: React.CSSProperties
-  theme?: any
-}) {
-  const bg =
-    data ||
-    (((theme as any).colors[
-      "editor.lineHighlightBackground"
-    ] ||
-      (theme as any).colors[
-        "editor.selectionHighlightBackground"
-      ]) as string)
-  return (
-    <div
-      style={{
-        ...style,
-        background: bg,
-        // cursor: "pointer",
-      }}
-      // onClick={_ => alert("clicked")}
-    >
-      {children}
-    </div>
-  )
-}
-
-function Label({
-  children,
-  data,
-  style,
-  theme,
-}: {
-  data: any
-  children: React.ReactNode
-  style?: React.CSSProperties
-  theme?: any
-}) {
-  const bg = ((theme as any).colors[
-    "editor.lineHighlightBackground"
-  ] ||
-    (theme as any).colors[
-      "editor.selectionHighlightBackground"
-    ]) as string
-  const [hover, setHover] = React.useState(false)
-
-  return (
-    <div
-      style={{
-        ...style,
-        background: hover ? bg : undefined,
-      }}
-      onMouseEnter={() => setHover(true)}
-      onMouseLeave={() => setHover(false)}
-    >
-      {children}
-      <div
-        style={{
-          position: "absolute",
-          right: 0,
-          paddingRight: 16,
-          display: hover ? "block" : "none",
-          opacity: 0.7,
-        }}
-      >
-        {data?.children || data}
-      </div>
-    </div>
-  )
-}
-
-export const annotationsMap: Record<
-  string,
-  CodeAnnotation["Component"]
-> = {
-  box: Box,
-  bg: Background,
-  label: Label,
-  link: CodeLink,
-}
+import { annotationsMap } from "../client/annotations"
 
 export function getAnnotationsFromMetastring(
   options: Record<string, string>
