+++ packages/mdx/src/plugin/code-files-reducer.tsx
