+++ packages/mdx/src/plugin/code.ts
@@ -7,15 +7,8 @@ import {
   NodeInfo,
   splitChildren,
 } from "./unist-utils"
-import React from "react"
-import {
-  CodeSpring,
-  CodeStep,
-} from "@code-hike/smooth-code"
-import {
-  EditorSpring,
-  EditorProps,
-} from "@code-hike/mini-editor"
+import { CodeStep } from "@code-hike/smooth-code"
+import { EditorProps } from "@code-hike/mini-editor"
 import {
   getAnnotationsFromMetastring,
   extractAnnotationsFromCode,
@@ -38,26 +31,6 @@ export async function transformCodeNodes(
     }
   )
 }
-
-export function Code(props: EditorProps) {
-  if (
-    !props.southPanel &&
-    props.files.length === 1 &&
-    !props.files[0].name
-  ) {
-    return (
-      <CodeSpring
-        className="ch-code"
-        config={props.codeConfig}
-        step={props.files[0]}
-      />
-    )
-  } else {
-    return <EditorSpring {...props} />
-  }
-}
-
-// remark:
 
 export function isEditorNode(node: Node) {
   return (
